value_from_component: Always fetch capability components lazily.
When an array element is extracted from an array for printing, GDB
calls value_from_component which assumes it can copy a subset of bytes
from the containing value (e.g. the entire array).  Howver, that doesn't
work for capability types since a value can only hold a single tag and
not a set of tags for a larger range.  For now, when a component is
a capability, always create a lazy value which will be resolved via
target_read_capability.

Fixes #55