morello: Rework morello_write_pc.
Make morello_write_pc a bit closer to RISC-V in that it tries to use
the existing bounds from PCC if the new PC is in bounds.

Unlike RISC-V this retains the previous behavior of falling back to
broad bounds if the PCC would be untagged otherwise.  This does not
attempt to derive anything from CSP since CSP should not be executable
and CSP's bounds should not cover any text segments.