put_frame_register: Store new tag value before other bytes.
regcache::cooked_write pushes the new register value to the target
(e.g. ptrace()), so the tag needs to be set before cooked_write is
called.

While here, simplify this case to always set the tag on tagged
registers clearing it to false if the from value is not tagged.