linux-nat: Better handle alternate ABI siginfo structures
Alternate ABI siginfo_t structures (such as 32-bit on a 64-bit host)
have a different size than the native siginfo_t.  Previously
linux_xfer_siginfo conflated the sizes and would permit, for example,
reading stack garbage in the on-stack version.  Instead, add a new
low_siginfo_size hook to return the size of the current inferior's
siginfo_t and use this size for various bounds checks on the xfer
as well as to allocate a buffer for the converted siginfo_t.