cheri: linux: Fix dynamic debug sizes for CHERI
The debugger reads structures in target memory to determine
the list of loaded sharead libraries. Make sure that the
target specific description of these structures matches
what musl actually puts in memory for CHERI enabled systems.

This also makes sure that we match FreeBSDs definition of
these structures for CHERI.