cheri: dwarf2: Revert gdb/dwarf2/expr.c to upstream gdb version
There is a small diff that does not seem to be justified by
CHERI changes. It very much look like an eariler buggy version
from some snapshot between gdb-11 and gdb-12 that got imported
and dragged along ever since.

At least one of the diffs is positively buggy and makes gdb
crash with a CHERI out of bounds error when copying DWARF2 values.