gdb: Add one more hack to gdb/dwarf2/expr.c
Apparently, it is possible that the source for the copy
is smaller than the target and vice versa. Make sure that
we always restrict the copy to the size that is available
on both ends.

It is unclear (to me) what the correct fix for this might be
but the current version seems to work well and avoid gdb crashes.
So keep it for the time being.